[move] #6 ClickableVerticalRegionListInModal.kt 생성 후 이곳으로 ClickableVerticalRegionListInModal 함수 이동
